* Support decimal RunDuration, UI/UX and loader improvements
- Changed TripModel.RunDuration from int? to decimal? for fractional hour support; updated all references accordingly.
- Added ToTimeString extension to format decimal hours as "H:MM".
- TripDetailsCard now displays both running KM and formatted time.
- ManageTrip.razor: Added IsLoading flag and loader, improved fetch method for UI responsiveness, and added defensive dictionary access.
- GcItemCard.razor: EwayBill dialog now opens with improved sizing.
- EwayBillDialog.razor: Improved layout and padding.
- Customcss.css: Enhanced input adornment spacing.
- Minor formatting and whitespace fixes.

+++ ArmsModels/BaseModels/Operations/TripModel.cs
@@ -64,7 +64,7 @@ namespace ArmsModels.BaseModels
         public string Destination { get; set; }
         public int? Distance { get; set; }
         public int? RunKM { get; set; }
-        public int? RunDuration { get; set; }
+        public decimal? RunDuration { get; set; }
         public decimal? Fuel { get; set; }
         public decimal? Freight { get; set; }
         public decimal? Expenses { get; set; }
